More logic to set planned_end_date.
Not done yet, but getting there. We're now looking up the analysis by
the invocation/external ID, getting the list of the status updates,
checking to see if there are at least one of them in the Running state,
and checking to see if the planned_end_date is set.

Next up is actually adding the logic to compute and set the
planned_end_date.